Adding more examples and ignoring compiled files

+++ compile.rb
@@ -4,11 +4,19 @@ require 'haml'
 require 'sass'
 require 'coffee-script'
 
-`haml #{file}.haml #{file}.html`
-`sass #{file}.sass:#{file}.css`
-t = CoffeeScript.compile File.read("#{file}.coffee")
-File.open("#{file}.js", "w") do |f|
-  f.puts t
+if File.exists? "#{file}.haml"
+  `haml #{file}.haml #{file}.html`
+end
+
+if File.exists? "#{file}.sass"
+  `sass #{file}.sass:#{file}.css`
+end
+
+if File.exists? "#{file}.coffee"
+  t = CoffeeScript.compile File.read("#{file}.coffee")
+  File.open("#{file}.js", "w") do |f|
+    f.puts t
+  end
 end
 
 `open #{file}.html`
